pyproject.toml
[tool.poetry]
name = "mailto-scraper"
version = "0.1.0"
description = "Email scraper from a list of websites"
authors = ["Pere Pasamonte <perepc@gmail.com>"]
readme = "README.md"
packages = [{include = "mailto_scraper", from = "src"},
            {include = "shopify_searcher", from = "src"},
]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.0"
email-validator = "^2.2.0"
googlesearch-python = "^1.3.0"
pandas = "^2.2.3"
langchain = "^0.3.19"
dotenv = "^0.9.9"
langchain-community = "^0.3.18"
pydantic = "^2.10.6"
google-api-python-client = "^2.162.0"
google-auth = "^2.38.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 